Conflicts:
	.gitignore
	build.gradle
	svnkit/src/main/java/org/tmatesoft/svn/core/SVNException.java
	svnkit/src/main/java/org/tmatesoft/svn/core/SVNURL.java
	svnkit/src/main/java/org/tmatesoft/svn/core/internal/io/fs/FSRepository.java

+++ svnkit/src/main/java/org/tmatesoft/svn/core/internal/wc2/ng/SvnNgRevert.java
@@ -1,30 +1,17 @@
 package org.tmatesoft.svn.core.internal.wc2.ng;
 
-import java.io.File;
-import java.util.Collection;
-import java.util.Map;
-import java.util.Set;
-
-import org.tmatesoft.svn.core.SVNDepth;
-import org.tmatesoft.svn.core.SVNErrorCode;
-import org.tmatesoft.svn.core.SVNErrorMessage;
-import org.tmatesoft.svn.core.SVNException;
-import org.tmatesoft.svn.core.SVNNodeKind;
-import org.tmatesoft.svn.core.SVNProperties;
-import org.tmatesoft.svn.core.SVNProperty;
+import org.tmatesoft.svn.core.*;
+import org.tmatesoft.svn.core.internal.db.SVNSqlJetDb;
 import org.tmatesoft.svn.core.internal.util.SVNSkel;
 import org.tmatesoft.svn.core.internal.wc.SVNErrorManager;
 import org.tmatesoft.svn.core.internal.wc.SVNEventFactory;
 import org.tmatesoft.svn.core.internal.wc.SVNFileType;
 import org.tmatesoft.svn.core.internal.wc.SVNFileUtil;
 import org.tmatesoft.svn.core.internal.wc17.SVNWCContext;
-import org.tmatesoft.svn.core.internal.wc17.db.ISVNWCDb;
+import org.tmatesoft.svn.core.internal.wc17.db.*;
 import org.tmatesoft.svn.core.internal.wc17.db.ISVNWCDb.SVNWCDbKind;
 import org.tmatesoft.svn.core.internal.wc17.db.ISVNWCDb.SVNWCDbStatus;
-import org.tmatesoft.svn.core.internal.wc17.db.SVNWCDb;
-import org.tmatesoft.svn.core.internal.wc17.db.Structure;
 import org.tmatesoft.svn.core.internal.wc17.db.StructureFields.NodeInfo;
-import org.tmatesoft.svn.core.internal.wc17.db.SvnWcDbRevert;
 import org.tmatesoft.svn.core.internal.wc17.db.SvnWcDbRevert.RevertInfo;
 import org.tmatesoft.svn.core.wc.ISVNEventHandler;
 import org.tmatesoft.svn.core.wc.SVNEvent;
@@ -32,6 +19,9 @@ import org.tmatesoft.svn.core.wc.SVNEventAction;
 import org.tmatesoft.svn.core.wc2.SvnRevert;
 import org.tmatesoft.svn.core.wc2.SvnTarget;
 import org.tmatesoft.svn.util.SVNLogType;
+
+import java.io.File;
+import java.util.*;
 
 public class SvnNgRevert extends SvnNgOperationRunner<Void, SvnRevert> {
 
@@ -125,14 +115,80 @@ public class SvnNgRevert extends SvnNgOperationRunner<Void, SvnRevert> {
             getWcContext().writeCheck(localAbsPath);
         }
         try {
+
+            //we should detect that the copy was modified here
+            //after opRevert() call we won't be able to do that
+            final Set<File> modifiedCopiesThatShouldBePreserved = new HashSet<File>();
+            populateModifiedCopiesThatShouldBePreserved(localAbsPath, wcRoot, modifiedCopiesThatShouldBePreserved);
+
             getWcContext().getDb().opRevert(localAbsPath, depth);
-            restore(getWcContext(), localAbsPath, depth, useCommitTimes, getWcContext().getEventHandler());
+            restore(getWcContext(), localAbsPath, depth, useCommitTimes, getWcContext().getEventHandler(), modifiedCopiesThatShouldBePreserved);
         } finally {
             SvnWcDbRevert.dropRevertList(getWcContext(), localAbsPath);
         }
     }
 
+    private void populateModifiedCopiesThatShouldBePreserved(File localAbsPath, File wcRoot, Set<File> modifiedCopiesThatShouldBePreserved) throws SVNException {
+        if (!getOperation().isPreserveModifiedCopies()) {
+            return;
+        }
+        try {
+            final ISVNWCDb.WCDbInfo wcDbInfo = getWcContext().getDb().readInfo(localAbsPath, ISVNWCDb.WCDbInfo.InfoField.checksum,
+                    ISVNWCDb.WCDbInfo.InfoField.originalRootUrl, ISVNWCDb.WCDbInfo.InfoField.propsMod,
+                    ISVNWCDb.WCDbInfo.InfoField.lastModTime, ISVNWCDb.WCDbInfo.InfoField.translatedSize,
+                    ISVNWCDb.WCDbInfo.InfoField.kind);
+
+            final SVNNodeKind nodeKind = SVNFileType.getNodeKind(SVNFileType.getType(localAbsPath));
+            if (nodeKind == SVNNodeKind.NONE ) {
+                //there's nothing to preserve
+            } else if (nodeKind == SVNNodeKind.DIR) {
+                if (wcDbInfo.kind == SVNWCDbKind.Dir) {
+                    final List<File> children = getWcContext().getChildrenOfWorkingNode(localAbsPath, false);
+                    for (File child : children) {
+                        populateModifiedCopiesThatShouldBePreserved(child, wcRoot, modifiedCopiesThatShouldBePreserved);
+                    }
+                }
+            } else {
+                if (wcDbInfo.kind != SVNWCDbKind.File) {
+                    //obstruction, nothing to preserve
+                    return;
+                }
+
+
+                if (wcDbInfo.propsMod) {
+                    modifiedCopiesThatShouldBePreserved.add(localAbsPath);
+                    return;
+                }
+                if (wcDbInfo.originalRootUrl != null && wcDbInfo.checksum != null) {
+                    //now we are sure that the file is copied
+                    long fileSize = SVNFileUtil.getFileLength(localAbsPath);
+                    long fileTime = SVNFileUtil.getFileLastModified(localAbsPath);
+
+                    if (wcDbInfo.translatedSize != -1 && wcDbInfo.lastModTime != 0 &&
+                            wcDbInfo.translatedSize == fileSize && wcDbInfo.lastModTime == fileTime) {
+                        //not modified
+                    } else {
+                        SVNWCDb db = (SVNWCDb) getWcContext().getDb();
+                        SVNWCDb.DirParsedInfo dirParsedInfo = db.parseDir(wcRoot, SVNSqlJetDb.Mode.ReadOnly);
+                        File pristineFileName = SvnWcDbPristines.getPristineFileName(dirParsedInfo.wcDbDir.getWCRoot(), wcDbInfo.checksum, false);
+                        if (getWcContext().compareAndVerify(localAbsPath, pristineFileName, true, false, false)) {
+                            modifiedCopiesThatShouldBePreserved.add(localAbsPath);
+                        }
+                    }
+                }
+            }
+        } catch (SVNException e) {
+            if (e.getErrorMessage().getErrorCode() != SVNErrorCode.ENTRY_NOT_FOUND && e.getErrorMessage().getErrorCode() != SVNErrorCode.WC_PATH_NOT_FOUND) {
+                throw e;
+            }
+        }
+    }
+
     public static void restore(SVNWCContext context, File localAbsPath, SVNDepth depth, boolean useCommitTimes, ISVNEventHandler notifier) throws SVNException {
+        restore(context, localAbsPath, depth, useCommitTimes, notifier, Collections.<File>emptySet());
+    }
+
+    public static void restore(SVNWCContext context, File localAbsPath, SVNDepth depth, boolean useCommitTimes, ISVNEventHandler notifier, Set<File> modifiedCopiesThatShouldBePreserved) throws SVNException {
         context.checkCancelled();
         
         Structure<RevertInfo> revertInfo = SvnWcDbRevert.readRevertInfo(context, localAbsPath);
@@ -183,10 +239,12 @@ public class SvnNgRevert extends SvnNgOperationRunner<Void, SvnRevert> {
         
         if (revertInfo.is(RevertInfo.copiedHere)) {
             if (revertInfo.get(RevertInfo.kind) == SVNWCDbKind.File && onDisk == SVNNodeKind.FILE) {
-                SVNFileUtil.deleteFile(localAbsPath);
+                if (!modifiedCopiesThatShouldBePreserved.contains(localAbsPath)) {
+                    SVNFileUtil.deleteFile(localAbsPath);
+                }
                 onDisk = SVNNodeKind.NONE;
             } else if (revertInfo.get(RevertInfo.kind) == SVNWCDbKind.Dir && onDisk == SVNNodeKind.DIR) {
-                boolean removed = restoreCopiedDirectory(context, localAbsPath, true);
+                boolean removed = restoreCopiedDirectory(context, localAbsPath, true, modifiedCopiesThatShouldBePreserved);
                 if (removed) {
                     onDisk = SVNNodeKind.NONE;
                 }
@@ -212,8 +270,8 @@ public class SvnNgRevert extends SvnNgOperationRunner<Void, SvnRevert> {
                     SVNFileUtil.deleteFile(localAbsPath);
                     onDisk = SVNNodeKind.NONE;                    
                 } else {
-                    long lastModified = localAbsPath.lastModified();
-                    long size = localAbsPath.length();
+                    long lastModified = SVNFileUtil.getFileLastModified(localAbsPath);
+                    long size = SVNFileUtil.getFileLength(localAbsPath);
                     if (recordedSize != -1
                             && recordedTime != 0
                             && recordedSize == size 
@@ -286,19 +344,19 @@ public class SvnNgRevert extends SvnNgOperationRunner<Void, SvnRevert> {
         }
         
         if (depth == SVNDepth.INFINITY && kind == SVNWCDbKind.Dir) {
-            restoreCopiedDirectory(context, localAbsPath, false);
+            restoreCopiedDirectory(context, localAbsPath, false, modifiedCopiesThatShouldBePreserved);
             
             Set<String> children = ((SVNWCDb) context.getDb()).getChildrenOfWorkingNode(localAbsPath);
             for (String childName : children) {
                 File childAbsPath = SVNFileUtil.createFilePath(localAbsPath, childName);
-                restore(context, childAbsPath, depth, useCommitTimes, notifier);
+                restore(context, childAbsPath, depth, useCommitTimes, notifier, modifiedCopiesThatShouldBePreserved);
             }
         }
         
         SvnWcDbRevert.notifyRevert(context, localAbsPath, notifier);
     }
 
-    private static boolean restoreCopiedDirectory(SVNWCContext context, File localAbsPath, boolean removeSelf) throws SVNException {
+    private static boolean restoreCopiedDirectory(SVNWCContext context, File localAbsPath, boolean removeSelf, Set<File> modifiedCopiesThatShouldBePreserved) throws SVNException {
         boolean selfRemoved = false;
         
         Map<File, SVNWCDbKind> children = SvnWcDbRevert.readRevertCopiedChildren(context, localAbsPath);
@@ -313,7 +371,12 @@ public class SvnNgRevert extends SvnNgOperationRunner<Void, SvnRevert> {
             if (childFileType != SVNFileType.FILE && childFileType != SVNFileType.SYMLINK) {
                 continue;
             }
-            SVNFileUtil.deleteFile(child);
+            if (!modifiedCopiesThatShouldBePreserved.contains(child)) {
+                SVNFileUtil.deleteFile(child);
+            } else {
+                selfRemoved = false;
+                removeSelf = false;
+            }
         }
 
         for (File child : children.keySet()) {
@@ -322,7 +385,12 @@ public class SvnNgRevert extends SvnNgOperationRunner<Void, SvnRevert> {
             if (childKind != SVNWCDbKind.Dir) {
                 continue;
             }
-            SVNFileUtil.deleteFile(child);
+            if (!modifiedCopiesThatShouldBePreserved.contains(child)) {
+                SVNFileUtil.deleteFile(child);
+            } else {
+                selfRemoved = false;
+                removeSelf = false;
+            }
         }
         
         if (removeSelf) {
